requirements.txt
flask
google-api-python-client
google-auth
google-auth-oauthlib
openai>=1.0.0
